+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - hnefatafl-client: social preview and hide letters.
+- hnefatafl-client: draw a bordered box around the board.
 
 ## [4.3.0] - 2025-11-24